Main Function Complete
All required functions are all been added.
Restriction is yet to be fully Implemented.
Sub functions are yet to be added:
-Task Punishments
-Bonus Reward for Level up
-more

+++ TaskSaga/Models/ExpReward.cs
@@ -7,6 +7,6 @@ namespace TaskSaga.Models
     internal class ExpReward
     {
         public string Name { get; set; }
-        public int Exp { get; set; }
+        public double Exp { get; set; }
     }
 }
