+++ TaskSaga/Models/Stats.cs
@@ -7,7 +7,7 @@ namespace TaskSaga.Models
     internal class Stats
     {
         public int Level { get; set; }
-        public int Exp { get; set; }
+        public double Exp { get; set; }
         public int ExpCap { get; set; }
     }
 }
